add websocket event to update order and set status print, pay and taken

+++ backend/src/orders/orders.controller.ts
@@ -87,8 +87,9 @@ export class OrdersController {
   @Delete(':id')
   @ApiOkResponse({ type: OrderEntity })
   delete(@Param('id') id: string,
-    @GetCurrentUser('role') role: string) {
-    return this.ordersService.delete({ id }, role);
+    @GetCurrentUser('role') role: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.delete({ id }, role, userId);
   }
 
   @Post('/detail/:orderDetailId/print')
@@ -103,8 +104,10 @@ export class OrdersController {
   @Post('/detail/:orderDetailId/cancel-print')
   @Roles(Role.Admin, Role.Operator)
   @HttpCode(200)
-  cancelMarkPrinted(@Param('orderDetailId') id: string,) {
-    return this.ordersService.cancelStatus({ type: CancelType.PRINT, orderDetailId: id });
+  cancelMarkPrinted(
+    @Param('orderDetailId') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.PRINT, orderDetailId: id, userId });
   }
 
   @Post('/:id/pay')
@@ -118,8 +121,9 @@ export class OrdersController {
   @Post('/:id/cancel-pay')
   @Roles(Role.Admin, Role.Administrasi)
   @HttpCode(200)
-  cancelMarkPay(@Param('id') id: string) {
-    return this.ordersService.cancelStatus({ type: CancelType.PAY, orderId: id });
+  cancelMarkPay(@Param('id') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.PAY, orderId: id, userId });
   }
 
   @Post('/:id/taken')
@@ -133,7 +137,8 @@ export class OrdersController {
   @Post('/:id/cancel-taken')
   @Roles(Role.Admin, Role.Administrasi)
   @HttpCode(200)
-  cancelMarkTaken(@Param('id') id: string) {
-    return this.ordersService.cancelStatus({ type: CancelType.TAKEN, orderId: id });
+  cancelMarkTaken(@Param('id') id: string,
+    @GetCurrentUserId() userId: number) {
+    return this.ordersService.cancelStatus({ type: CancelType.TAKEN, orderId: id, userId });
   }
 }
